refactor(tests): split test_triage_repair.py under the 300-LOC gate
The bloat Stop-hook flagged a NEW crossing: test_triage_repair.py reached 326
lines (limit 300). It is a new crossing rather than a ratchet of an existing
baseline entry, so the pre-commit anti-ratchet hook did not block it and it
landed with #399.

Split along the file's own section boundaries rather than taking an ADR
exception — the two halves answer different questions:

  test_triage_repair.py        (185) scan + apply: what the repair DOES
  test_triage_repair_safety.py (180) minimal rewrite, refusals, lock ordering:
                                     what the repair must never DESTROY
  _triage_repair_helpers.py     (49) shared fixtures, underscore-prefixed so
                                     pytest does not collect it (_sweep_helpers
                                     precedent)

Landmine worth recording: shared/tests/ and shared/scripts/ BOTH contain a real
`tools` package. Putting shared/tests on sys.path ahead of shared/scripts makes
`tools` resolve to shared/tests/tools, and `from tools.triage_repair import ...`
fails with ModuleNotFoundError. So shared/scripts is inserted at position 0
unconditionally (a membership guard would leave a pre-existing lower-priority
entry in place) and the sibling-helper dir is APPENDED at the end.

Re-verified non-vacuous after the move: forcing LF in the rewrite still fails
test_crlf_tracked_log_is_not_reflowed_to_lf, and moving the scan back outside
the lock still fails test_apply_scans_inside_the_lock.

Run-ID: iterate-2026-07-18-bloat-split-triage-repair-tests
